+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 ### Changed
-- Minor README.md refinements
+- Additional handling for OpenAI's strict schema specification
 
 
 ## [0.2.9] - 2025-06-24